engine: add support for local Git repositories
