+++ engine/testdata/engine/checktypes_trivy.json
@@ -4,8 +4,16 @@
             "name": "vulcan-trivy",
             "description": "Check based on aquasec/trivy",
             "image": "vulcansec/vulcan-trivy:edge",
-            "assets": ["DockerImage"],
+            "assets": [
+                "DockerImage",
+                "GitRepository"
+            ],
             "options": {
+                "git_checks": {
+                    "vuln": true,
+                    "secret": true,
+                    "config": true
+                },
                 "image_checks": {
                     "vuln": true,
                     "secret": true,
